+++ lib/grape/dsl/callbacks.rb
@@ -44,7 +44,7 @@ module Grape
           namespace_stackable(:afters, block)
         end
 
-        # Allows you to specify a block of code to be executed after every
+        # Allows you to specify a something that will always be executed after a call
         # API call. Unlike the `after` block, this code will run even on
         # unsuccesful requests
         # @example
@@ -52,7 +52,7 @@ module Grape
         #     before do
         #       ApiLogger.start
         #     end
-        #     ensure_block do
+        #     finally do
         #       ApiLogger.close
         #     end
         #   end
@@ -60,7 +60,7 @@ module Grape
         # This will make sure that the ApiLogger is opened and close around every
         # request
         # @param  ensured_block [Proc] The block to be executed after every api_call
-        def ensure_block(&block)
+        def finally(&block)
           namespace_stackable(:finallies, block)
         end
       end