+++ lib/grape/endpoint.rb
@@ -272,7 +272,6 @@ module Grape
           # The Body commonly is an Array of Strings, the application instance itself, or a File-like object
           response_object = file || [body]
 
-
           [status, header, response_object]
         ensure
           # An error on the finallies should not prevent a succesful request from executing